feat: Remove Build rollout dialog (#1297)
* Remove Build rollout dialog

* chore: Remove unneeded package-lock (#1308)

Remove unneeded package-lock

* fix: Console error on add source (#1311)

Change onClick function to take boolean

* Consolidate DiffDialog and DiffSection, add logic to inactivate compare versions button

---------
