--- ui/src/components/BuildRolloutDialog/index.ts
@@ -1,1 +1,0 @@
-export { BuildRolloutDialog } from "./BuildRolloutDialog";
