+++ ui/src/components/DiffDialog/diff-section.module.scss
