+++ ui/src/components/DiffDialog/get-rendered-config-data.ts
